+++ src/core/use-cases/create-todo/create-todo.interface.ts
@@ -1,5 +1,5 @@
 import { Todo } from '../../entities';
 
 export interface ICreateTodoUseCase {
-  create(dto: Todo): string;
+  create(dto: Todo): Promise<string>;
 }