+++ src/core/use-cases/create-todo/create-todo.spec.ts
@@ -1,18 +1,48 @@
 import { assert, expect } from 'chai';
 import sinon from 'sinon';
 import { Todo, TodoStatus } from '../../entities';
+import { JsonPlaceHolderUser } from '../../types';
 import { CreateTodoUseCase } from '..';
 
 describe('UseCase - CreateTodoUseCase', () => {
   describe('create', () => {
-    it('Should create todo with succeffully', () => {
+    it('Should create todo with succeffully', async () => {
       const todo_repository_mock = {
         create: sinon.fake.returns(undefined),
         list: sinon.fake.returns(undefined),
       };
 
+      const mock_user: JsonPlaceHolderUser = {
+        company: {
+          bs: 'yolo',
+          catchPhrase: 'yolo',
+          name: 'yolo',
+        },
+        email: 'yolo@yolo.com',
+        id: 1,
+        name: 'yolo',
+        phone: 'yolo',
+        username: 'yolo',
+        website: 'yolo',
+        address: {
+          city: 'yolo',
+          geo: {
+            lat: 'yolo',
+            lng: 'yolo',
+          },
+          street: 'yolo',
+          suite: 'yolo',
+          zipcode: 'yolo',
+        },
+      };
+
+      const json_place_holder_integration_mock = {
+        getUser: sinon.fake.resolves(mock_user),
+      };
+
       const use_case_context = {
         todo_repository: todo_repository_mock,
+        json_place_holder_integration: json_place_holder_integration_mock,
       };
 
       // @ts-ignore
@@ -24,7 +54,7 @@ describe('UseCase - CreateTodoUseCase', () => {
         user: 'yolo',
       };
 
-      const result = use_case.create(dto);
+      const result = await use_case.create(dto);
       expect(result).to.be.equals('Todo created!');
 
       assert(todo_repository_mock.create.calledOnceWith(new Todo(dto)));