+++ src/infra/integrations/http/index.ts
@@ -1,0 +1,5 @@
+export * from './client';
+export * from './interfaces';
+export * from './enum';
+export * from './jsonplaceholder/jsonplaceholder';
+export * from './jsonplaceholder/jsonplaceholder.response';