+++ src/interface/cli/create-todo/create-todo.ts
@@ -4,8 +4,8 @@ import { CliLine, CliResult, ICreateTodoCommand } from '..';
 export class CreateTodoCommand implements ICreateTodoCommand {
   constructor(private create_todo_use_case: ICreateTodoUseCase) {}
 
-  cmd(cmd: CliLine): CliResult<string> {
-    const result = this.create_todo_use_case.create(cmd.line);
+  async cmd(cmd: CliLine): Promise<CliResult<string>> {
+    const result = await this.create_todo_use_case.create(cmd.line);
 
     return {
       result,