+++ src/main/container/container.ts
@@ -6,6 +6,10 @@ import {
 } from '../../core/use-cases';
 import { TodoRepository } from '../../infra/repositories';
 import { UseCaseContext } from '.';
+import {
+  HttpClient,
+  JsonPlaceHolderIntegration,
+} from '../../infra/integrations/http';
 
 export class Container {
   readonly list_todo_use_case: IListTodoUseCase;
@@ -13,8 +17,13 @@ export class Container {
   readonly create_todo_use_case: ICreateTodoUseCase;
 
   constructor() {
+    const client_http = new HttpClient();
+
     const use_case_context: UseCaseContext = {
       todo_repository: new TodoRepository([]),
+      json_place_holder_integration: new JsonPlaceHolderIntegration(
+        client_http
+      ),
     };
 
     this.list_todo_use_case = new ListTodoUseCase(use_case_context);