+++ src/main/container/container.type.ts
@@ -1,5 +1,6 @@
-import { ITodoRepository } from '../../core/ports';
+import { IJsonPlaceHolderIntegration, ITodoRepository } from '../../core/ports';
 
 export type UseCaseContext = {
   todo_repository: ITodoRepository;
+  json_place_holder_integration: IJsonPlaceHolderIntegration;
 };